split Geography and Fit classes; missing graphics utils;
