+++ python/cv20.py
@@ -8,8 +8,7 @@ Created on Thu Jul  2 09:04:11 2020
 @author: jsibert
 """
 
-import pandas as pd
-from datetime import date, datetime, timedelta
+"""
 from dateutil.relativedelta import relativedelta
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
@@ -30,7 +29,11 @@ import re
 import statistics
 
 plt.style.use('file:///home/jsibert/.config/matplotlib/john.mplstyle')
-import js_covid as cv
+"""
+
+#import js_covid as cv
+from js_covid import *
+from Geography import *
 
 eps = 1e-5
 
@@ -225,714 +228,8 @@ def add_signature(fig,url_line):
     fig.text(1.0,0.020,url_line+' ', ha='right',va='bottom', fontsize=8,alpha=0.25)#,color='red')
 
 # -----------  class definitions--------------       
-
-class Geography:
-
-#   def __init__(self,name,enclosed_by,code):
-    def __init__(self,**kwargs):
-        self.gtype = None
-        self.name = kwargs.get('name')
-        self.enclosed_by = kwargs.get('enclosed_by')
-        self.code = kwargs.get('code')
-        self.population = None
-        self.source = None
-        if self.name is None:
-            self.moniker = None
-        else:
-            self.moniker = str(self.name+self.code)
-            self.moniker =  self.moniker.replace(' ','_',5) 
-    #   self.TMB_fit = None
-    #   self.ADMB_fit = None
-        if self.moniker is None:
-            self.dat_file = None
-        else:
-            self.dat_file = cv.dat_path+self.moniker+'.dat'
-        self.updated = None
-        self.date0 = None
-        self.ntime = None
-        self.date = None
-        self.cases = None
-        self.deaths = None
-        self.pdate = None # date for plotting on x-axis
-        
-    def print(self):
-        print(self)
-        
-        
-    def print_metadata(self):
-        print('gtype:',self.gtype)
-        print('gname:',self.name)
-        print('genclosed_by:',self.enclosed_by)
-        print('gcode:',self.code)
-        print('gpopulation:',self.population)
-        print('gsource:',self.source)
-        print('gmoniker:',self.moniker)
-        print('gdat_file:',self.dat_file)
-        print('gupdated:',self.updated)
-        print('gdate0:',self.date0)
-        print('gntime:',self.ntime)
-
-    def print_data(self):
-        self.get_pdate()
-    #   if (self.deaths != None):
-        if (len(self.deaths) > 0):
-            print('date','cases','deaths','pdate')
-        else:
-            print('date','cases','pdate')
-        for k in range(0,len(self.date)):
-        #   if (self.deaths != None):
-            if (len(self.deaths) > 0):
-                print(self.date[k],self.cases[k],self.deaths[k],self.pdate[k])
-            else:
-                print(self.date[k],self.cases[k],self.pdate[k])
-            
-    def to_DataFrame(self):
-        self.get_pdate()
-        Gdf = pd.DataFrame(columns=('date','pdate','cases','deaths'))
-        Gdf['date'] = self.date
-        Gdf['pdate'] = self.pdate
-        Gdf['cases'] = self.cases
-        Gdf['deaths'] = self.deaths
-        return(Gdf)
-
-    def get_pdate(self):
-        if (self.pdate != None):
-            return(self.pdate)
-
-        else:
-            self.pdate = []
-            for d in self.date:
-                self.pdate.append(mdates.date2num(datetime.strptime(d,'%Y-%m-%d').date()))
-            return(self.pdate)
-
-    def get_county_pop(self):
-        """
-        
-        """
-
-        dat = cv.population_dat
-        if (dat.empty):
-            print('Reading',cv.census_data_path)
-            cv.population_dat = pd.read_csv(cv.census_data_path,header=0,comment='#')
-            dat = cv.population_dat
-    #   else:
-    #       print('Using current "dat" object')
-
-        state_filter = dat['state'].isin([self.enclosed_by])
-        county_filter = dat['county'].isin([self.name])
-    #   COUNTY_filter = (dat['COUNTY']>0)
-        County_rows = state_filter & county_filter #& COUNTY_filter
-        try:
-            population = int(pd.to_numeric(dat[County_rows]['population'].values))
-        except:
-            print('get_county_pop() failed for:')
-            print(self.name, self.enclosed_by,self.code) 
-            population = 1
-        return(population)   
-        
-    def read_nyt_data(self,gtype='county'):
-        self.gtype = gtype
-        mtime = 0.0
-        if (gtype == 'county'):
-            dat = cv.nyt_county_dat
-            if (dat.empty):
-               cspath = cv.NYT_counties
-               print('Reading',cspath)
-               cv.nyt_county_dat = pd.read_csv(cspath,header=0)
-               dat = cv.nyt_county_dat
-               mtime = os.path.getmtime(cspath)
-        else:
-            sys.exit('No data for',gtype)
-        
-    #   dat = pd.read_csv(cspath,header=0)
-
-    #   mtime = os.path.getmtime(cspath)
-        dtime = datetime.fromtimestamp(mtime)
-        self.updated = str(dtime.date())
-        self.population = self.get_county_pop()
-    
-        state_filter = dat['state'].isin([self.enclosed_by])
-        county_filter = dat['county'].isin([self.name])
-        County_rows = state_filter & county_filter
-        if (len(County_rows) < 1):
-            sys.exit(' * * * no records found for',self.name,self.surrounded_by)
-
-   #    tmp = dat[County_rows]['date'].map(Strptime)
-   #    self.pdate  = np.array(mdates.date2num(tmp))
-        self.cases  = np.array(dat[County_rows]['cases'])
-        self.deaths = np.array(dat[County_rows]['deaths'])
-        self.date   = np.array(dat[County_rows]['date'])
-        self.date0 = self.date[0]
-        self.ntime = len(self.date)
-        self.source = 'New York Times, https://github.com/nytimes/covid-19-data.git.'
-        
-    def write_dat_file(self):
-        print(len(self.date),'records found for',self.name,self.enclosed_by)
-        O = open(self.dat_file,'w')
-        O.write('# county\n %s\n'%(self.moniker))
-        O.write('# updated from https://github.com/nytimes/covid-19-data.git\n')
-        O.write(' %s\n'%self.updated)
-        O.write('# population (N0)\n %10d\n'%self.population)
-        O.write('# date zero\n %s\n'%self.date0)
-        ntime = len(self.date)-1
-        O.write('# ntime\n %4d\n'%ntime)
-        O.write('#%6s %5s\n'%('cases','deaths'))
-        for r in range(0,len(self.date)):
-            O.write('%7d %6d\n'%(self.cases[r],self.deaths[r]))
-
-    def read_BCHA_data(self,gtype='hsa'):
-        self.gtype = gtype
-        cspath = cv.BCHA_path
-
-        dat = pd.read_csv(cspath,header=0)
-
-        mtime = os.path.getmtime(cspath)
-        dtime = datetime.fromtimestamp(mtime)
-        self.updated = str(dtime.date())
-        self.population = None
-   #    self.source = 'British Columbia Centre for Disease Control'\
-        self.source = 'Government of British Columbia'\
-                      ' www.bccdc.ca/Health-Info-Site/Documents/' 
-
-        rdates = pd.DatetimeIndex(dat["Reported_Date"],normalize=True)
-        self.date = pd.date_range(rdates[0],rdates[len(rdates)-1],normalize=True,freq='D')
-        self.date = self.date.strftime('%Y-%m-%d') # NYT style
-        self.date0 = self.date[0]
-        self.ntime = len(self.date)
-        self.get_pdate()
-
-
-        daily_cases = pd.Series([0]*self.ntime)
-
-        HA_filter = dat["HA"].isin([self.name])
-        dat["rep_date"] = rdates
-        for i in range(0,len(self.date)):
-        #   date_mask = dat["Reported_Date"].isin([self.date[i]])
-            date_mask = dat["rep_date"].isin([self.date[i]])
-            mask = HA_filter & date_mask
-            daily_cases[i] = len(dat[mask.values])
-
-        self.cases = daily_cases.cumsum()
-        
-
-
-    def plot_prevalence(self,yscale='linear', per_capita=False, delta_ts=True,
-                        window=[7], plot_dt = False, cumulative = True,
-                        show_order_date = True,
-                        annotation = True, signature = False, 
-                        save = True, dashboard = False):
-        
-        """ 
-        Plots cases and deaths vs calendar date 
-
-        scale: select linear of log scale 'linear'
-        per_capita: plot numbers per 1000 (see mult)  False
-        delta_ts: plot daily new cases True
-        window: plot moving agerage window [11]
-        plot_dt: plot initial doubling time slopes on log scale False
-        annotations: add title and acknowledgements True
-        save : save plot as file
-        """
-        mult = 1000
-
-        firstDate = datetime.strptime(self.date[0],'%Y-%m-%d')
-        orderDate = mdates.date2num(cv.CAOrderDate)
-
-        if (self.deaths is None):
-            nax = 1
-        else:
-            nax = 2
-    
-        fig, pax = plt.subplots(nax,1,figsize=(6.5,nax*2.25))
-        ax = [None]*2
-        if (nax == 1):
-            ax[0] = pax
-        else:
-            ax=pax
-
-        if (per_capita):
-            ax[0].set_ylabel('Daily Cases'+' per '+str(mult))
-            if (nax > 1):
-                ax[1].set_ylabel('Daily Deaths'+' per '+str(mult))
-        else:
-            ax[0].set_ylabel('Daily Cases')
-            if (nax > 1):
-                ax[1].set_ylabel('Daily Deaths')
-    
-        ax2 = []
-        for a in range(0,nax):
-            self.make_date_axis(ax[a],firstDate)
-            ax[a].set_yscale(yscale)
-            if (cumulative):
-                ax2.append(ax[a].twinx())
-                ax2[a].set_ylabel('Cumulative')
-        
-        if (per_capita):
-            cases =  mult*self.cases/self.population + eps
-            if (nax > 1):
-                deaths =  mult*self.deaths/self.population + eps
-            else:
-                deaths =  self.deaths
-        else:
-            cases  =  self.cases
-            deaths =  self.deaths
-        
-        nn = self.ntime-1
-        max_cases = cases[nn]
-        if (self.deaths is None):
-            max_deaths = 0.0
-        else:
-            max_deaths = deaths[nn]
-
-        Date = self.get_pdate()
-
-        delta_cases = np.diff(cases)
-        ax[0].bar(Date[1:], delta_cases)
-
-        if (max_cases < 1.0):
-            none_reported(ax[0],'Cases')
-        else:
-            for w in range(0,len(window)):
-                adc = pd.Series(delta_cases).rolling(window=window[w]).mean()
-                ax[0].plot(Date[1:],adc,linewidth=2)
-                mark_ends(ax[0],Date[1:],adc, str(window[w])+'da','r')
-                max_adc = adc.max()
-        
-            if (cumulative):
-                ax2[0].plot(Date, cases,alpha=0.5, linewidth=1)#,label=cc)
-                mark_ends(ax2[0],Date,cases,r'$\Sigma$C','r')
-
-            if ((yscale == 'log') & (plot_dt)):
-                ax[0] = self.plot_dtslopes(ax[0])
-                
-        if (nax > 1):
-            if (cumulative):
-                ax2[1].plot(Date, deaths,alpha=0.5,linewidth=1)#,label=cc)
-                mark_ends(ax2[1],Date,deaths,r'$\Sigma$D','r')
-             
-
-            if (max_deaths > 0.0):
-                delta_deaths = np.diff(deaths)
-                ax[1].bar(Date[1:],delta_deaths)
-                for w in range(0,len(window)):
-                    add = pd.Series(delta_deaths).rolling(window=window[w]).mean()
-                    ax[1].plot(Date[1:],add,linewidth=2)
-                    mark_ends(ax[1],Date[1:],add, str(window[w])+'da','r')
-                    max_add = add.max()
-            else:
-                none_reported(ax[1],'Deaths')
-    
-    #   Adjust length of y axis
-        ax[0].set_ylim(0.0,1.2*max_adc) #SD_lim(delta_cases,3.0)[1]) #ax[a].get_ylim()[1])
-        tx = prop_scale(ax[0].get_xlim(), 0.5)
-        ty = prop_scale(ax[0].get_ylim(), 0.95)
-        note = '{0:,} Cases'.format(int(max_cases))
-        ax[0].text(tx,ty,note ,ha='center',va='top',fontsize=10)
-        if (max_deaths > 0.0):
-            ax[1].set_ylim(0.0,1.2*max_add) #SD_lim(delta_deaths,3.0)[1]) #ax[a].get_ylim()[1])
-            tx = prop_scale(ax[1].get_xlim(), 0.5)
-            ty = prop_scale(ax[1].get_ylim(), 0.95)
-            note = '{0:,} Deaths'.format(int(max_deaths))
-            ax[1].text(tx,ty,note ,ha='center',va='top',fontsize=10)
-
-        for a in range(0,nax):
-            if (delta_ts):
-                if (cumulative):
-                    ax2[a].set_ylim(0,ax2[a].get_ylim()[1])
-        #   Newsome's shelter in place order
-            if (show_order_date):
-                ax[a].plot((orderDate,orderDate),
-                           (ax[a].get_ylim()[0], ax[a].get_ylim()[1]),
-                           color='black', linewidth=3,alpha=0.5)
-        #   ax[a].legend()
-    
-        if (annotation):
-            if (self.gtype == 'county'):
-                gname = 'County'
-            else:
-                gname = 'Region'
-            title = 'Covid-19 Prevalence in '+self.name+' '+gname+', '+ self.enclosed_by
-            fig.text(0.5,1.0,title ,ha='center',va='top')
-        #   fig.text(0.0,0.0,' Data source: '+ self.source ,
-        #            ha='left',va='bottom', fontsize=8)
-    
-        #   mtime = os.path.getmtime(cv.NYT_home+'us-counties.csv')
-        #   dtime = datetime.fromtimestamp(mtime)
-        #   fig.text(1.0,0.0,'Updated '+str(dtime.date())+' ', ha='right',va='bottom', fontsize=8)
-            add_data_source(fig,self.source)
-
-        if (signature):
-        #   by_line = 'Graphics by John Sibert'
-        #   url_line = 'https://github.com/johnrsibert/SIR-Models/tree/master/PlotsToShare'
-        #   fig.text(0.0,0.020,' '+by_line, ha='left',va='bottom', fontsize=8,alpha=0.25)#,color='red')
-        #   fig.text(1.0,0.020,url_line+' ', ha='right',va='bottom', fontsize=8,alpha=0.25)#,color='red')
-            add_signature(fig,'https://github.com/johnrsibert/SIR-Models/tree/master/PlotsToShare')
-    
-        if (dashboard):
-        #   out_img = BytesIO()
-        #   plt.savefig(out_img, format='png')
-        #   out_img.seek(0)  # rewind file
-        #   encoded = base64.b64encode(out_img.read()).decode("ascii").replace("\n", "")
-        #   return "data:image/png;base64,{}".format(encoded)
-            gfile = cv.graphics_path+'test.png'
-            
-            print('saving fig for dash:',gfile)
-            plt.savefig(gfile,format='png',dpi=300)
-            plt.close()
-            print('saved')
-
-        else:
-            if save:
-                gfile = cv.graphics_path+self.moniker+'_prevalence.png'
-                plt.savefig(gfile,dpi=300)
-                plt.show(False)
-                plt.pause(3)
-                plt.close()
-                print('plot saved as',gfile)
-            else:
-                plt.show()
-
-
-    def make_date_axis(self, ax, first_prev_date = None):
-        if (first_prev_date is None):
-            firstDate = mdates.date2num(cv.FirstNYTDate)
-        else:
-            firstDate = first_prev_date
-        
-        lastDate  = mdates.date2num(cv.EndOfTime)
-    #   print('firstDate,lastDate:',firstDate,lastDate)
-        ax.set_xlim([firstDate,lastDate])
-        ax.xaxis.set_major_formatter(mdates.DateFormatter("%b"))
-        ax.xaxis.set_major_locator(mdates.MonthLocator())
-        ax.xaxis.set_minor_locator(mdates.DayLocator())
-
-
-    def plot_dtslopes(self, ax, threshold = 2, dt = [1,2,4,8]):
-        """
-        Superimpose exponential growth slope lines for different doubling times
-        ax: axisis on which to draw slopes lines
-        threshold: number of cases used to start slopes
-        dt: representative doubling times in days
-        """
-        k0 = 0
-        # find frist date in shich cases exceeds threshold
-        for k in range(0, self.ntime):
-            if (self.cases[k] >= threshold):
-                k0 = k
-                break
-
-        d0 = self.pdate[k0]
-        c0 = self.cases[k0]
-        sl = np.log(2.0)/dt
-        xrange = ax.get_xlim()
-        yrange = [25,ax.get_ylim()[1]]
-        for i in range(0,len(dt)):
-            y = c0 + np.exp(sl[i]*(d0-xrange[0]))
-            ax.plot([d0,xrange[1]],[c0,y],color='black',linewidth=1)
-            c = ax.get_lines()[-1].get_color()
-            mark_ends(ax,xrange,y,str(dt[i])+' da','r')
-
-        return(ax)
-
-    def plot_per_capita_curvature(self,mult = 1000,save=False):
-        pc_cases = self.cases/self.population * mult
-        
-        fig, ax = plt.subplots(1,figsize=(6.5,4.5))
-        self.make_date_axis(ax)
-        
-        grad = np.gradient(pc_cases)
-    #   d1 = np.diff(pc_cases)
-    #   print(len(d1))
-    #   grad = np.diff(d1)
-    #   print(len(grad))
-
-    #   ax.plot(self.get_pdate()[2:],grad)
-        ax.plot(self.get_pdate(),grad)
-        print(min(grad),max(grad))
-
-        plt.show()
-
-
-
-class Fit(Geography):
-
-    def __init__(self,fit_file,**kwargs):
-        super().__init__(**kwargs)
-        pn = fit_file #cv.fit_path+fit_file
-        filename, extension = os.path.splitext(pn)
-        if (extension == '.RData'):
-            self.fit_type = 'TMB'
-            tfit=pyreadr.read_r(pn)
-        else:
-            sys.exit('class Fit  not yet implemented on '+extension+' files.')
-
-        head, tail = os.path.split(pn)
-        self.moniker = os.path.splitext(tail)[0]
-        self.diag = tfit['diag']
-        self.md = tfit['meta']
-        self.ests = tfit['ests']
-        if (cv.pyreadr_kludge):
-        #   print('keys',tfit.keys())
-        #   print(tfit['diag'].columns)
-        #   print(len(tfit['diag'].columns))
-            insert = len(tfit['diag'].columns)
-        #   print('head:',head,'tail:',tail,tail[0])
-        #   print('filename',filename,'extension',extension)
-            csv_file = filename + '_stderror.csv'
-        #   print('reading',csv_file)
-            stderror = pd.read_csv(csv_file,header=0)
-        #   print(stderror['logbeta'],stderror['logmu'])
-        #   SElogbeta = stderror['logbeta']
-        #   SElogmu = stderror['logmu']
-            self.diag.insert(insert,'SElogbeta',stderror['logbeta'])
-            self.diag.insert(insert+1,'SElogmu',stderror['logmu'])
-        #   print(self.diag)
-
-        self.date0 = self.get_metadata_item('Date0')
-        self.ntime = int(self.get_metadata_item('ntime'))
-    #   Date0 = self.get_metadata_item('Date0')
-    #   print('Date0:',Date0)
-
-    def print_metadata(self):
-        super().print_metadata()
-        print('type:',self.fit_type)
-
-    def get_metadata_item(self,mdname):
-        r = self.md['names'].isin([mdname])
-        return(self.md.data[r].values[0])
-
-    def get_estimate_item(self, ename):
-        r = self.ests['names'].isin([ename])
-        if (r.any() == True):
-            return(float(self.ests.est[r]))
-        else:
-            return(None)
-    
-    def get_est_or_init(self,name):
-        v = self.get_estimate_item(name) 
-        if (isNaN(v)):
-            v = self.get_initpar_item(name)
-            return(v)
-        else:
-            return(v)
-    
-    def get_initpar_item(self,pname):
-        r = self.ests['names'].isin([pname])
-        if (r.any() == True):
-            return(float(ests.init[r]))
-        else:
-            return(None)
-       
-    def plot(self,logscale=True, per_capita=False, delta_ts=False,
-             npl = 4, save = True, show_median = False):
-        """ 
-        """
-    
-        firstDate = mdates.date2num(cv.FirstNYTDate)
-        orderDate = mdates.date2num(cv.CAOrderDate)
-        lastDate  = mdates.date2num(cv.EndOfTime)
-    
-        date_list = pd.date_range(start=firstDate,end=lastDate)
-        date_lim = [date_list[0],date_list[len(date_list)-1]]
-        plt.rcParams['lines.linewidth'] = 1
-        plt.rcParams["scatter.marker"] = '+'
-        plt.rcParams["lines.markersize"] = 6
-        prefix = ''
-        if (logscale):
-            prefix = 'ln '
-             
-        fig, ax = plt.subplots(npl,1,figsize=(9.0,(npl)*3.0))
-        if (per_capita):
-            ax[0].set_ylabel(prefix+'Cases'+' per '+str(mult))
-            ax[1].set_ylabel(prefix+'Deaths'+' per '+str(mult))
-        else:
-            ax[0].set_ylabel(prefix+'Cases')
-            ax[1].set_ylabel(prefix+'Deaths')
-        if (npl > 2):
-            ax[2].set_ylabel(r'$\ln\ \beta\ (da^{-1})$')
-        if (npl > 3):
-            ax[3].set_ylabel(r'$\ln\ \mu\ (da^{-1})$')
-    
-    
-        for a in range(0,len(ax)):
-            self.make_date_axis(ax[a])
-    
-        Date0 = self.get_metadata_item('Date0')
-        Date0 = datetime.strptime(Date0,'%Y-%m-%d')
-        pdate = []
-        for t in range(0,len(self.diag.index)):
-            pdate.append(mdates.date2num(Date0 + timedelta(days=t)))
-    
-        obsI = self.diag['log_obs_cases']
-        preI = self.diag['log_pred_cases']
-        obsD = self.diag['log_obs_deaths']
-        preD = self.diag['log_pred_deaths']
-        SElogbeta = self.diag['SElogbeta']
-        SElogmu   = self.diag['SElogmu']
-        sigma_logC = np.exp(self.get_est_or_init('logsigma_logC'))
-        sigma_logD = np.exp(self.get_est_or_init('logsigma_logD'))
-        sigma_logbeta = self.get_est_or_init('logsigma_logbeta')
-        sigma_logmu = np.exp(self.get_est_or_init('logsigma_logmu'))
-        if (logscale):
-            ax[0].set_ylim(0.0,1.2*max(obsI))
-            ax[0].scatter(pdate,obsI,marker = '.',s=16,c='r')
-            ax[0].plot(pdate,preI)
-
-        else:
-            ax[0].set_ylim(0.0,1.2*max(np.exp(obsI)))
-            ax[0].scatter(pdate,np.exp(obsI),marker = '.',s=16,c='r')
-            ax[0].plot(pdate,np.exp(preI))
-        plot_error(ax[0],pdate,obsI,sigma_logC,logscale)
-        tx = prop_scale(ax[0].get_xlim(), 0.05)
-        ty = prop_scale(ax[0].get_ylim(), 0.90)
-        sigstr = '%s = %.3g'%('$\sigma_I$',sigma_logC)
-        ax[0].text(tx,ty,sigstr, ha='left',va='center',fontsize=10)
-    
-        if (logscale):
-            ax[1].set_ylim(0.0,1.2*max(obsD))
-            ax[1].scatter(pdate,obsD,marker = '.',s=16,c='r')
-            ax[1].plot(pdate,preD)
-        else:
-            ax[1].set_ylim(0.0,1.2*max(np.exp(obsD)))
-            ax[1].scatter(pdate,np.exp(obsD),marker = '.',s=16, c='r')
-            ax[1].plot(pdate,np.exp(preD))
-
-        plot_error(ax[1],pdate,obsD,sigma_logD,logscale)
-        tx = prop_scale(ax[1].get_xlim(), 0.05)
-        ty = prop_scale(ax[1].get_ylim(), 0.90)
-        sigstr = '%s = %.3g'%('$\sigma_D$',sigma_logD)
-        ax[1].text(tx,ty,sigstr, ha='left',va='center',fontsize=10)
-    
-        if (npl > 2):
-            log_beta = self.diag['logbeta']
-        #   log_beta_ticks = [ 1.01978144,0.32663426,-0.36651292,
-        #                     -1.0596601,-1.75280728,-2.44595446,
-        #                     -3.13910164,-3.83224882,-4.525396  ]
-        #   print(min(log_beta_ticks),max(log_beta_ticks))
-        #   print(0.8*min(log_beta_ticks),1.2*max(log_beta_ticks))
-        #   ax[2].set_ylim((1.2*min(log_beta)),1.2*max(log_beta))
-        #   ax[2].set_yticks(log_beta_ticks)
-            sigstr = '%s = %.3g'%('$\sigma_\\beta$',sigma_logbeta)
-            tx = prop_scale(ax[2].get_xlim(), 0.05)
-            ty = prop_scale(ax[2].get_ylim(), 0.90)
-            ax[2].plot(pdate,log_beta)
-        #   plot_error(ax[2],pdate,log_beta,sigma_logbeta,logscale=True)
-            plot_error(ax[2],pdate,log_beta,SElogbeta,logscale=True)
-            ax[2].text(tx,ty,sigstr, ha='left',va='center',fontsize=10)
-
-            if (show_median):
-                med = median(np.exp(log_beta))
-                logmed = np.log(med)
-                medstr = '%s = %.3g'%('$\\tilde{\\beta}$',med)
-                ax[2].text(ax[2].get_xlim()[0],logmed,medstr,ha='left',va='bottom',fontsize=10)
-                ax[2].plot(ax[2].get_xlim(),[logmed,logmed])
-
-        #   increase frequcncy of tick marks
-            start, end = ax[2].get_ylim()
-            dtick = (end - start)/5
-            ax[2].set_yticks(np.arange(start, end, dtick))
-
-        #   finagle doubling time axis at same scale as beta
-            add_doubling_time = False
-            if (add_doubling_time):
-                dtax = ax[2].twinx()
-                dtax.set_ylim(ax[2].get_ylim())
-                dtax.grid(False,axis='y') # omit grid lines
-                dtax.set_ylabel('Doubling Time (da)')
-            #   render now to get the tick positions and labels
-            #   fig.show()
-            #   fig.canvas.draw()
-                y2_ticks = dtax.get_yticks()
-                labels = dtax.get_yticklabels()
-                labels[0] = ''
-                for i in range(1,len(y2_ticks)):
-                    y2_ticks[i] = np.log(2)/np.exp(y2_ticks[i])
-                    labels[i] = round(float(y2_ticks[i]),2)
-                      
-                dtax.tick_params(length=0)
-                dtax.set_yticklabels(labels)
-            #   fig.show()
-            #   fig.canvas.draw()
-
-        if (npl > 3):
-            logmu = self.diag['logmu']
-        #   ymin = min(self.diag['logmu'])-2.5*sigma_logmu
-        #   ax[3].set_ylim(ymin,1.2*max(self.diag['logmu']))
-        #   print(min(logmu),max(logmu))
-        #   print(1.2*min(logmu),0.8*max(logmu))
-            ax[3].set_ylim((1.2*min(logmu),0.8*max(logmu)))
-
-            sigstr = '%s = %.3g'%('$\sigma_\\mu$',sigma_logmu)
-            tx = prop_scale(ax[3].get_xlim(), 0.05)
-            ty = prop_scale(ax[3].get_ylim(), 0.90)
-            ax[3].plot(ax[2].get_xlim(),[0.0,0.0],color='0.2',linestyle='--')
-            ax[3].plot(pdate,self.diag['logmu'])
-        #   plot_error(ax[3],pdate,logmu,sigma_logmu,logscale=True)
-            plot_error(ax[3],pdate,logmu,SElogmu,logscale=True)
-            ax[3].text(tx,ty,sigstr, ha='left',va='center',fontsize=10)
-
-            if (show_median):
-                med = median(np.exp(self.diag['logmu']))
-                logmed = np.log(med)
-                medstr = '%s = %.3g'%('$\\tilde{\\mu}$',med)
-                ax[3].text(ax[3].get_xlim()[0],logmed,medstr,ha='left',va='bottom',fontsize=10)
-                ax[3].plot(ax[3].get_xlim(),[logmed,logmed])
-    
-        title = self.moniker #self.name+' County, '+self.enclosed_by
-        fig.text(0.5,0.95,title ,ha='center',va='bottom')
-
-        if save:
-            gfile = cv.graphics_path+self.moniker+'_'+self.fit_type+'_estimates'
-            if (not logscale):
-                gfile = cv.graphics_path+self.moniker+'_'+self.fit_type+'_estimates_a'
-            fig.savefig(gfile+'.png',dpi=300)
-            plt.show(False)
-            print('plot saved as',gfile)
-            plt.pause(3)
-            plt.close()
-        else:
-            plt.show(True)
-
-    def plot_CMR(self, save = True):
-        npl = 3
-        fig, ax = plt.subplots(npl,1,figsize=(9.0,(npl)*3.0))
-        ax[0].set_ylabel('Empirical CMR')
-        ax[1].set_ylabel('Estimated CMR')
-        ax[2].set_ylabel(r'$\mu\ (da^{-1})$')
- 
-        for a in range(0,len(ax)):
-            self.make_date_axis(ax[a])
-    
-        Date0 = self.get_metadata_item('Date0')
-        Date0 = datetime.strptime(Date0,'%Y-%m-%d')
-        pdate = []
-        for t in range(0,len(self.diag.index)):
-            pdate.append(mdates.date2num(Date0 + timedelta(days=t)))
-        
-    #   self.get_pdate()
-        obsI = self.diag['obs_cases']
-        obsD = self.diag['obs_deaths']
-        obsCMR = obsD/obsI
-        estI = np.exp(self.diag['log_pred_cases'])
-        estD = np.exp(self.diag['log_pred_deaths'])
-        estCMR = estD/estI 
-        mu   = np.exp(self.diag['logmu'])
-
-        ax[0].plot(pdate,obsCMR)
-        ax[1].set_ylim(ax[0].get_ylim())
-        ax[1].plot(pdate,estCMR)
-        ax[2].plot(pdate,mu)
-
-        fig.text(0.5,0.95,'Case Mortality Ratio' ,ha='center',va='bottom')
-        if (save):
-            gfile = cv.graphics_path+self.moniker+'_CMR.png'
-            fig.savefig(gfile)
-            plt.show(False)
-            print('plot saved as',gfile)
-            plt.pause(3)
-        else:
-            plt.show()
-
+# moved to Geography.py
+# moved to fit.py
 # ----------- end of class definitions--------------       
 
 def make_SD_tab(Gfile='top30.csv',save=True):
@@ -1713,29 +1010,9 @@ def log_norm_cfr():
 
 #
 
-#log_norm_cfr()
-
-#mean = d_cfr.mean[0]
-#sigma = math.sqrt(d_cfr.variance[0])
-#print(mean,sigma)
-#x = np.linspace(mean - 3.0*sigma, mean + 3.0*sigma, 50) 
-#pdf =stats.norm.pdf(x, mean, sigma)
-#pdfsum = sum(pdf)
-#fig, ax = plt.subplots(1,figsize=(6.5,6.5))
-#ax.plot(x,pdf/pdf.sum())
-#ax.plot((mean,mean),ax.get_ylim())
-#plt.show()
 
 # -------------------------------------------------
 
-#tgeog = Geography(name='Los Angeles',enclosed_by='California',code='CA')
-#tgeog = Geography(name='New York City',enclosed_by='New York',code='NY')
-#tgeog = Geography(name='Santa Clara',enclosed_by='California',code='CA')
-#tgeog = Geography(name='Harris',enclosed_by='Texas',code='TX')
-#tgeog.read_nyt_data('county')
-#tgeog.print_metadata()
-#tgeog.print_data()
-#tgeog.plot_prevalence(save=False,cumulative=False, show_order_date=False,signature=True)
 
 #cv.fit_path = cv.fit_path+'constrainID/'
 #tfit = Fit(cv.fit_path+'CookIL.RData') #'Los Angeles','California','CA','ADMB')
@@ -1941,3 +1218,13 @@ def make_cfr_histo_ts(nG = 100,save=True):
     plt.show()#block=False)
 
 #make_cfr_histo_ts()
+
+
+tgeog = Geography(name='Los Angeles',enclosed_by='California',code='CA')
+#tgeog = Geography(name='New York City',enclosed_by='New York',code='NY')
+#tgeog = Geography(name='Santa Clara',enclosed_by='California',code='CA')
+#tgeog = Geography(name='Harris',enclosed_by='Texas',code='TX')
+tgeog.read_nyt_data('county')
+tgeog.print_metadata()
+tgeog.print_data()
+tgeog.plot_prevalence(save=False,cumulative=False, show_order_date=False,signature=True)
